Lift tests so they can be reused by homestore implementation once working.
